Fix race condition causing flaky errors
Having a mutable set in a population, which was shared among multiple
different threads (each agent) and mutated by each was a bad idea.
Making it an immutable set, which is a var, means that each change
happens as a transaction, and then the value gets updated. You may get
dirty reads or lost writes, but we have already established that we
don't care about 100% transactional correctness.

Authored-by: Julian Zucker <julian.zucker@gmail.com>